refactor: difficulty settings
-Create utility function to handle create, get, and delete requests for difficulty settings via menu provider.
